Fix bug in polytope package, only clean spack directory if uberenv stuff is there

+++ scripts/devtools/tpl-manager.py
@@ -115,7 +115,8 @@ class SpheralTPL:
                 sexe(f"git -C {spack_dir} remote add origin {self.args.spack_url}")
                 sexe(f"git -C {spack_dir} fetch --depth=2 origin {spack_commit}")
                 sexe(f"git -C {spack_dir} checkout FETCH_HEAD")
-                if (self.args.clean):
+                uber_env_trash = os.path.join(spack_dir, "etc/spack/defaults/upstreams.yaml")
+                if (self.args.clean and os.path.exists(uber_env_trash)):
                     sexe(f"git -C {spack_dir} clean -df")
             else:
                 # Check commit hash of Spack repo
